refactor: Move create fields into create sub-app
The majority of these fields are only used by the create journey
so they are being moved into the relevant part of the app that is
responsible for them.
