+++ app/move/app/new/fields/common.assessment-checkboxes.js
