+++ app/move/app/new/fields/common.filter.js
