+++ app/move/app/new/fields/filter.prison-number.js
@@ -1,6 +1,6 @@
 const { cloneDeep } = require('lodash')
 
-const validators = require('../validators')
+const validators = require('../../../validators')
 
 const filter = require('./common.filter')
