+++ app/move/app/new/fields/time-due.js
@@ -1,5 +1,5 @@
-const { time: timeFormatter } = require('../../../common/formatters')
-const { datetime: datetimeValidator, after } = require('../validators')
+const { time: timeFormatter } = require('../../../../../common/formatters')
+const { datetime: datetimeValidator, after } = require('../../../validators')
 
 const timeDue = {
   id: 'time_due',
